creates a file for renderAsideGroups
