+++ public/js/views/renderAside/renderAsideCategory.js
@@ -1,0 +1,9 @@
+import app from "../../app.js";
+const renderAsideCategory = (category, group) => {
+  app.view.createElement(
+    "li",
+    category,
+    document.getElementById(`sidebar_${group}`)
+  );
+};
+export default renderAsideCategory;
